Enable auto-reconnect while logging in already (but lower the timeout)
Fixes login failures if the server takes a couple of seconds to start up
(as can be the case when the database needs to be created from scratch).

The lower reconnect timeout is so that it's more obvious if the server
takes unusually long to start or otherwise fails to respond.