Remove unused void* params from NetComm functions and messages
Some of these already had no effect at all. Most ended up in
plNetCommReplyMsg::param, which was never read anywhere. Usually, the
parameters were just set to nullptr.

This does mean that there will be no way to trace the origin of
plNetCommReplyMsgs anymore. This can be re-implemented in the future if
necessary.